requirements.txt
openai
pydantic
python-pptx
tenacity
rich